+++ changelog.md
@@ -2,3 +2,5 @@
 * Changed road barrier texture & top texture of barriers posts.
 * Changed barriers texture mapping (removed `barrier_post.png`)
 * Removed barriers `side_tall` models (identical and therefore replaced by `side` models)
+* Potted trees now use logs as texture instead of small logs
+* Fixed wrong potted mangrove tree model